fix var size bug

+++ alexb/apps/config_test/main.cc
@@ -60,7 +60,7 @@ class TestConfig : public NovaProgram, public ProgramConsole
 
       ConfigProtocol *service_config = new ConfigProtocol(alloc_cap(ConfigProtocol::CAP_SERVER_PT + hip->cpu_desc_count()));
 
-      unsigned id;
+      unsigned short id;
       return (!service_config->start_config(*utcb, id, config));
     }
